+++ packages/analytics-browser/src/config.ts
@@ -66,6 +66,9 @@ export class BrowserConfig extends Config implements IBrowserConfig {
   constructor(apiKey: string, userId?: string, options?: BrowserOptions) {
     const defaultConfig = getDefaultConfig();
     super({
+      flushIntervalMillis: 1000,
+      flushMaxRetries: 5,
+      flushQueueSize: 30,
       ...options,
       apiKey,
       storageProvider: options?.storageProvider ?? defaultConfig.storageProvider,
@@ -186,15 +189,21 @@ export const createFlexibleStorage = async <T>(options: BrowserOptions): Promise
   return storage;
 };
 
-export const createEventsStorage = async (overrides?: BrowserOptions): Promise<Storage<Event[]>> => {
-  let eventsStorage = overrides?.storageProvider;
-  if (!eventsStorage || !(await eventsStorage.isEnabled())) {
-    eventsStorage = new LocalStorage();
-    if (!(await eventsStorage.isEnabled())) {
-      eventsStorage = new MemoryStorage();
+export const createEventsStorage = async (overrides?: BrowserOptions): Promise<Storage<Event[]> | undefined> => {
+  const hasStorageProviderProperty = overrides && Object.prototype.hasOwnProperty.call(overrides, 'storageProvider');
+  // If storageProperty is explicitly undefined `{ storageProperty: undefined }`
+  // then storageProvider is undefined
+  // If storageProvider is implicitly undefined `{ }`
+  // then storageProvider is LocalStorage
+  // Otherwise storageProvider is overriden
+  if (!hasStorageProviderProperty || overrides.storageProvider) {
+    for (const storage of [overrides?.storageProvider, new LocalStorage<Event[]>()]) {
+      if (storage && (await storage.isEnabled())) {
+        return storage;
+      }
     }
   }
-  return eventsStorage;
+  return undefined;
 };
 
 export const createDeviceId = (idFromCookies?: string, idFromOptions?: string, idFromQueryParams?: string) => {