+++ packages/analytics-core/src/config.ts
@@ -19,12 +19,11 @@ import {
 import { Logger } from './logger';
 
 export const getDefaultConfig = () => ({
-  flushMaxRetries: 5,
-  flushQueueSize: 10,
-  flushIntervalMillis: 1000,
+  flushMaxRetries: 12,
+  flushQueueSize: 200,
+  flushIntervalMillis: 10000,
   logLevel: LogLevel.Warn,
   loggerProvider: new Logger(),
-  saveEvents: true,
   optOut: false,
   serverUrl: AMPLITUDE_SERVER_URL,
   serverZone: ServerZone.US,
@@ -40,11 +39,10 @@ export class Config implements IConfig {
   logLevel: LogLevel;
   minIdLength?: number;
   plan?: Plan;
-  saveEvents: boolean;
   serverUrl: string | undefined;
   serverZone?: ServerZone;
   transportProvider: Transport;
-  storageProvider: Storage<Event[]>;
+  storageProvider?: Storage<Event[]>;
   useBatch: boolean;
 
   private _optOut = false;
@@ -66,7 +64,6 @@ export class Config implements IConfig {
     this.minIdLength = options.minIdLength;
     this.plan = options.plan;
     this.optOut = options.optOut ?? defaultConfig.optOut;
-    this.saveEvents = options.saveEvents ?? defaultConfig.saveEvents;
     this.serverUrl = options.serverUrl;
     this.serverZone = options.serverZone || defaultConfig.serverZone;
     this.storageProvider = options.storageProvider;