+++ packages/analytics-core/test/config.test.ts
@@ -19,16 +19,15 @@ describe('config', () => {
     });
     expect(config).toEqual({
       apiKey: 'apiKey',
-      flushIntervalMillis: 1000,
-      flushMaxRetries: 5,
-      flushQueueSize: 10,
+      flushIntervalMillis: 10000,
+      flushMaxRetries: 12,
+      flushQueueSize: 200,
       logLevel: LogLevel.Warn,
       loggerProvider: new Logger(),
       minIdLength: undefined,
       _optOut: false, // private for `optOut` getter/setter
       partnerId: undefined,
       plan: undefined,
-      saveEvents: true,
       serverUrl: 'https://api2.amplitude.com/2/httpapi',
       serverZone: 'US',
       storageProvider: defaultConfig.storageProvider,
@@ -45,16 +44,15 @@ describe('config', () => {
       logLevel: LogLevel.Verbose,
       optOut: true,
       plan: { version: '0' },
-      saveEvents: false,
       storageProvider: defaultConfig.storageProvider,
       transportProvider: defaultConfig.transportProvider,
       useBatch: true,
     });
     expect(config).toEqual({
       apiKey: 'apiKey',
-      flushIntervalMillis: 1000,
-      flushMaxRetries: 5,
-      flushQueueSize: 10,
+      flushIntervalMillis: 10000,
+      flushMaxRetries: 12,
+      flushQueueSize: 200,
       logLevel: LogLevel.Verbose,
       loggerProvider: new Logger(),
       minIdLength: undefined,
@@ -62,7 +60,6 @@ describe('config', () => {
       plan: {
         version: '0',
       },
-      saveEvents: false,
       serverUrl: 'https://api2.amplitude.com/batch',
       serverZone: 'US',
       storageProvider: defaultConfig.storageProvider,