+++ packages/analytics-core/test/timeline.test.ts
@@ -1,7 +1,6 @@
 import { Timeline } from '../src/timeline';
 import { Event, Plugin, PluginType } from '@amplitude/analytics-types';
-import { OPT_OUT_MESSAGE } from '../src/messages';
-import { useDefaultConfig } from './helpers/default';
+import { useDefaultConfig, promiseState } from './helpers/default';
 import { createTrackEvent } from '../src/utils/event-builder';
 
 describe('timeline', () => {
@@ -65,6 +64,8 @@ describe('timeline', () => {
     await timeline.register(before, config);
     await timeline.register(enrichment, config);
     await timeline.register(destination, config);
+    timeline.isReady = true;
+
     expect(beforeSetup).toHaveBeenCalledTimes(1);
     expect(enrichmentSetup).toHaveBeenCalledTimes(1);
     expect(destinationSetup).toHaveBeenCalledTimes(1);
@@ -73,18 +74,14 @@ describe('timeline', () => {
       event_type: `${id}:event_type`,
     });
     await Promise.all([
-      timeline.push(event(1), config).then(() => timeline.push(event(1.1), config)),
+      timeline.push(event(1)).then(() => timeline.push(event(1.1))),
+      timeline.push(event(2)).then(() => Promise.all([timeline.push(event(2.1)), timeline.push(event(2.2))])),
       timeline
-        .push(event(2), config)
-        .then(() => Promise.all([timeline.push(event(2.1), config), timeline.push(event(2.2), config)])),
-      timeline
-        .push(event(3), config)
+        .push(event(3))
         .then(() =>
           Promise.all([
-            timeline
-              .push(event(3.1), config)
-              .then(() => Promise.all([timeline.push(event(3.11), config), timeline.push(event(3.12), config)])),
-            timeline.push(event(3.2), config),
+            timeline.push(event(3.1)).then(() => Promise.all([timeline.push(event(3.11)), timeline.push(event(3.12))])),
+            timeline.push(event(3.2)),
           ]),
         ),
     ]);
@@ -100,18 +97,13 @@ describe('timeline', () => {
   });
 
   describe('push', () => {
-    test('should handle opt out', async () => {
+    test('should skip event processing when config is missing', async () => {
       const event = {
         event_type: 'hello',
       };
-      const config = useDefaultConfig();
-      config.optOut = true;
-      const results = await timeline.push(event, config);
-      expect(results).toEqual({
-        event,
-        code: 0,
-        message: OPT_OUT_MESSAGE,
-      });
+      const result = timeline.push(event);
+      expect(await promiseState(result)).toEqual('pending');
+      expect(timeline.queue.length).toBe(1);
     });
   });
 
@@ -152,10 +144,10 @@ describe('timeline', () => {
       };
       const config = useDefaultConfig();
       await timeline.register(plugin, config);
-      void timeline.push(createTrackEvent('a'), config);
-      void timeline.push(createTrackEvent('b'), config);
-      void timeline.push(createTrackEvent('c'), config);
-      void timeline.push(createTrackEvent('d'), config);
+      void timeline.push(createTrackEvent('a'));
+      void timeline.push(createTrackEvent('b'));
+      void timeline.push(createTrackEvent('c'));
+      void timeline.push(createTrackEvent('d'));
       expect(timeline.queue.length).toBe(4);
       await timeline.flush();
       expect(timeline.queue.length).toBe(0);