+++ packages/analytics-node/src/config.ts
@@ -1,10 +1,10 @@
-import { NodeOptions, NodeConfig as INodeConfig, Event } from '@amplitude/analytics-types';
-import { Config, MemoryStorage } from '@amplitude/analytics-core';
+import { NodeOptions, NodeConfig as INodeConfig } from '@amplitude/analytics-types';
+import { Config } from '@amplitude/analytics-core';
 import { Http } from './transports/http';
 
 export class NodeConfig extends Config implements INodeConfig {
   constructor(apiKey: string, options?: NodeOptions) {
-    const storageProvider = options?.storageProvider ?? new MemoryStorage<Event[]>();
+    const storageProvider = options?.storageProvider;
     const transportProvider = options?.transportProvider ?? new Http();
 
     super({