+++ packages/analytics-node/src/node-client.ts
@@ -4,15 +4,22 @@ import { Context } from './plugins/context';
 import { useNodeConfig } from './config';
 
 export class AmplitudeNode extends AmplitudeCore<NodeConfig> {
-  async init(apiKey: string, _userId?: string | undefined, options?: NodeOptions) {
+  async init(apiKey: string, options?: NodeOptions) {
     const nodeOptions = useNodeConfig(apiKey, {
       ...options,
     });
 
-    await super.init(undefined, undefined, nodeOptions);
+    await super._init(nodeOptions);
 
     await this.add(new Context());
     await this.add(new Destination());
+
+    // Set timeline ready for processing events
+    // Send existing events, which might be collected by track before init
+    this.timeline.isReady = true;
+    if (!this.config.optOut) {
+      this.timeline.scheduleApply(0);
+    }
   }
 }