+++ packages/analytics-react-native/test/config.test.ts
@@ -37,14 +37,13 @@ describe('config', () => {
         domain: '',
         flushIntervalMillis: 1000,
         flushMaxRetries: 5,
-        flushQueueSize: 10,
+        flushQueueSize: 30,
         loggerProvider: logger,
         logLevel: LogLevel.Warn,
         minIdLength: undefined,
         _optOut: false,
         partnerId: undefined,
         plan: undefined,
-        saveEvents: true,
         serverUrl: 'https://api2.amplitude.com/2/httpapi',
         serverZone: 'US',
         sessionManager,
@@ -100,14 +99,13 @@ describe('config', () => {
         domain: '',
         flushIntervalMillis: 1000,
         flushMaxRetries: 5,
-        flushQueueSize: 10,
+        flushQueueSize: 30,
         loggerProvider: logger,
         logLevel: LogLevel.Warn,
         minIdLength: undefined,
         _optOut: false,
         partnerId: undefined,
         plan: undefined,
-        saveEvents: true,
         serverUrl: 'https://api2.amplitude.com/2/httpapi',
         serverZone: 'US',
         sessionManager,
@@ -167,7 +165,7 @@ describe('config', () => {
         domain: '',
         flushIntervalMillis: 1000,
         flushMaxRetries: 5,
-        flushQueueSize: 10,
+        flushQueueSize: 30,
         loggerProvider: logger,
         logLevel: LogLevel.Warn,
         minIdLength: undefined,
@@ -283,18 +281,11 @@ describe('config', () => {
       expect(storage).toBeInstanceOf(LocalStorageModule.LocalStorage);
     });
 
-    test('should use memory', async () => {
-      const localStorageConstructor = jest.spyOn(LocalStorageModule, 'LocalStorage').mockReturnValueOnce({
-        isEnabled: async () => false,
-        get: async () => '',
-        set: async () => undefined,
-        remove: async () => undefined,
-        reset: async () => undefined,
-        getRaw: async () => undefined,
+    test('should return undefined storage', async () => {
+      const storage = await Config.createEventsStorage({
+        storageProvider: undefined,
       });
-      const storage = await Config.createEventsStorage();
-      expect(storage).toBeInstanceOf(core.MemoryStorage);
-      expect(localStorageConstructor).toHaveBeenCalledTimes(1);
+      expect(storage).toBe(undefined);
     });
   });