+++ packages/analytics-react-native/test/react-native-client.test.ts
@@ -194,6 +194,20 @@ describe('react-native-client', () => {
     });
   });
 
+  describe('reset', () => {
+    test('should reset user id and generate new device id config', async () => {
+      const client = new AmplitudeReactNative();
+      await client.init(API_KEY);
+      client.setUserId(USER_ID);
+      client.setDeviceId(DEVICE_ID);
+      expect(client.getUserId()).toBe(USER_ID);
+      expect(client.getDeviceId()).toBe(DEVICE_ID);
+      client.reset();
+      expect(client.getUserId()).toBe(undefined);
+      expect(client.getDeviceId()).not.toBe(DEVICE_ID);
+    });
+  });
+
   describe('getSessionId', () => {
     test('should get session id', async () => {
       const client = new AmplitudeReactNative();