+++ packages/analytics-node-test/CHANGELOG.md
@@ -3,6 +3,27 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.8](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-node-test@0.0.7...@amplitude/analytics-node-test@0.0.8) (2022-07-22)
+
+
+### Bug Fixes
+
+* adds error handling for invalid api ([#153](https://github.com/amplitude/Amplitude-TypeScript/issues/153)) ([c03f9d7](https://github.com/amplitude/Amplitude-TypeScript/commit/c03f9d7dad51e3026673dca31418a74591d79bbc))
+* missing tracked events before init issue ([#144](https://github.com/amplitude/Amplitude-TypeScript/issues/144)) ([60d0f68](https://github.com/amplitude/Amplitude-TypeScript/commit/60d0f6848087f7b8fc3c870d55489a238e841b26))
+* update default flush config for node ([#152](https://github.com/amplitude/Amplitude-TypeScript/issues/152)) ([2445dff](https://github.com/amplitude/Amplitude-TypeScript/commit/2445dff0842e7e0a2b7ee767ab926b5a93348214))
+
+
+
+
+
+## [0.0.7](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-node-test@0.0.6...@amplitude/analytics-node-test@0.0.7) (2022-07-15)
+
+**Note:** Version bump only for package @amplitude/analytics-node-test
+
+
+
+
+
 ## [0.0.6](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-node-test@0.0.5...@amplitude/analytics-node-test@0.0.6) (2022-07-13)
 
 **Note:** Version bump only for package @amplitude/analytics-node-test